move MetarParser to lib
